+++ src/components/NotificationBar/NotificationBar.jsx
@@ -1,13 +1,11 @@
-import { useState, useEffect } from 'react';
-import Container from '../Container/Container';
+import { useState, useEffect, lazy, Suspense } from 'react';
 import testLocalStorage from '../../utilities/test-local-storage';
 import './NotificationBar.scss';
-import CloseIcon from '../../styles/icons/cross.svg';
-import PropTypes from 'prop-types';
-import Content from './Notification.mdx';
 
-const version = '3';
-const localStorageIsEnabled = testLocalStorage() !== false;
+const MessageBar = lazy(() => import('./MessageBar'));
+
+export const version = '3';
+export const localStorageIsEnabled = testLocalStorage() !== false;
 
 const barDismissed = () => {
   if (localStorageIsEnabled) {
@@ -15,34 +13,6 @@ const barDismissed = () => {
   }
   return false;
 };
-
-MessageBar.propTypes = {
-  onClose: PropTypes.func,
-};
-
-function MessageBar(props) {
-  const close = () => {
-    localStorage.setItem('notification-dismissed', version);
-    props.onClose();
-  };
-  return (
-    <div className="notification-bar">
-      <Container className="notification-bar__inner">
-        <Content />
-        {localStorageIsEnabled ? (
-          <CloseIcon
-            aria-label="Dismiss"
-            className="notification-bar__close"
-            fill="#fff"
-            width={16}
-            onClick={close}
-            role="button"
-          />
-        ) : null}
-      </Container>
-    </div>
-  );
-}
 
 export default function NotificationBar() {
   // hide the bar in the beginning
@@ -66,6 +36,12 @@ export default function NotificationBar() {
     }
   }, [dismissed]);
   return (
-    <>{dismissed === false ? <MessageBar onClose={onClose} /> : undefined}</>
+    <>
+      {dismissed === false ? (
+        <Suspense fallback={<div />}>
+          <MessageBar onClose={onClose} />
+        </Suspense>
+      ) : undefined}
+    </>
   );
 }
